Fixed the first copy of Ease to be closed deleting all temporary directories.
Previous, Ease would essentially rm -rf /tmp/ease on each exit.
This was hardly safe, as multiple copies of the program could be using
that directory. Now, each process uses /tmp/ease/{PID}, and /tmp/ease is only
removed if it is empty (ie no copies of Ease are using it).