[general] Add generic Color and Gradient classes
Ease.Color

Color is a generic abstraction that supports Cairo, GDK, and
Clutter. all "color" usage in Ease should switch to this class
unless it is a "simple" usage - setting the background of a
Clutter stage, for example.

Ease.Gradient & Ease.GradientMode

Currently useless, will see support later in the form of slide
backgrounds and shapes.