[editor] Element type-specific inspector panes
All Element subclasses must now implement inspector_widget(),
which returns a widget that will be placed in the "element"
pane of the inspector when that element is selected.

Added inspector widgets for text and image elements.

Additionally, a virtual (but not abstract) method on Element
allows subclasses to define a set of toolbar items. However,
this is not currently used by EditorWindow, so all functionality
must be placed in the inspector widget.

Text-specific toolbar items have been removed, replaced by
the text element inspector widget.

The user can now set the alignment of text.