[general] Don't include unused media in saved documents
Saving was previous a "dumb" archive - just archive
everything in the temp directory. When media files
are "deleted" from the Document, they need to stick
around in the temp directory, in case the user reverts
the deletion but has moved the original file elsewhere.

However, undo isn't saved into .ease files, so there's
no need to include those extra media files. This commit
excludes any files that are not currently used
somewhere in the document.