[editor] Better handling of transition times in inspector.
When the user switches away from the "none" transition, a default
time will be set (if the time isn't already non-zero). The time
spin button is now disabled when the "none" transition is
selected.

When the user sets the slide to automatically advance, a default
time will also be set, once again if the time for that property
isn't already non-zero.

Fixed "None" transition playing the "Fade" transition if a
transition time greater than 0 was set.

Fixes second half of bug 627800